// src/wallet/model/adminRedemptionUpdateRequest.ts
/**
 * Zarban Wallet API
 * API for Zarban wallet services.
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: info@zarban.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AdminRedemptionUpdateRequest {
    /**
    * The status of the redemption request
    */
    'status': AdminRedemptionUpdateRequest.StatusEnum;
    /**
    * The Paya tracking code
    */
    'payaTrackingCode'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "AdminRedemptionUpdateRequest.StatusEnum"
        },
        {
            "name": "payaTrackingCode",
            "baseName": "payaTrackingCode",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AdminRedemptionUpdateRequest.attributeTypeMap;
    }
}

export namespace AdminRedemptionUpdateRequest {
    export enum StatusEnum {
        Approved = <any> 'Approved',
        Rejected = <any> 'Rejected',
        Completed = <any> 'Completed'
    }
}