// src/wallet/model/authTelegramRequest.ts
/**
 * Zarban Wallet API
 * API for Zarban wallet services.
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: info@zarban.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AuthTelegramRequest {
    /**
    * Telegram init data
    */
    'initdata': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "initdata",
            "baseName": "initdata",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AuthTelegramRequest.attributeTypeMap;
    }
}