// src/wallet/model/balance.ts
/**
 * Zarban Wallet API
 * API for Zarban wallet services.
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: info@zarban.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Coin } from './coin';
import { Currency } from './currency';

export class Balance {
    'locked': Currency;
    'coin': Coin;
    'balance': Currency;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "locked",
            "baseName": "locked",
            "type": "Currency"
        },
        {
            "name": "coin",
            "baseName": "coin",
            "type": "Coin"
        },
        {
            "name": "balance",
            "baseName": "balance",
            "type": "Currency"
        }    ];

    static getAttributeTypeMap() {
        return Balance.attributeTypeMap;
    }
}