// src/wallet/model/coin.ts
/**
 * Zarban Wallet API
 * API for Zarban wallet services.
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: info@zarban.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { BulletContent } from './bulletContent';
import { CoinConfig } from './coinConfig';
import { Network } from './network';
import { Symbol } from './symbol';

export class Coin {
    'config': CoinConfig;
    'symbol': Symbol;
    'name': { [key: string]: string; };
    'logoUri': string;
    'depositableNetworks': Array<Network>;
    'withdrawableNetworks': Array<Network>;
    'content': BulletContent;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "config",
            "baseName": "config",
            "type": "CoinConfig"
        },
        {
            "name": "symbol",
            "baseName": "symbol",
            "type": "Symbol"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "{ [key: string]: string; }"
        },
        {
            "name": "logoUri",
            "baseName": "logoUri",
            "type": "string"
        },
        {
            "name": "depositableNetworks",
            "baseName": "depositableNetworks",
            "type": "Array<Network>"
        },
        {
            "name": "withdrawableNetworks",
            "baseName": "withdrawableNetworks",
            "type": "Array<Network>"
        },
        {
            "name": "content",
            "baseName": "content",
            "type": "BulletContent"
        }    ];

    static getAttributeTypeMap() {
        return Coin.attributeTypeMap;
    }
}

export namespace Coin {
}