// src/wallet/model/coinConfig.ts
/**
 * Zarban Wallet API
 * API for Zarban wallet services.
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: info@zarban.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CoinConfig {
    'isTradeable': boolean;
    /**
    * Map of network to amount
    */
    'withdrawFees': { [key: string]: string; };
    /**
    * Map of network to amount
    */
    'minWithdrawal': { [key: string]: string; };
    'needsMemo': boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "isTradeable",
            "baseName": "isTradeable",
            "type": "boolean"
        },
        {
            "name": "withdrawFees",
            "baseName": "withdrawFees",
            "type": "{ [key: string]: string; }"
        },
        {
            "name": "minWithdrawal",
            "baseName": "minWithdrawal",
            "type": "{ [key: string]: string; }"
        },
        {
            "name": "needsMemo",
            "baseName": "needsMemo",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return CoinConfig.attributeTypeMap;
    }
}